+++ R/array-class.R
@@ -7,16 +7,21 @@ detectFronts.array <- function(x, method = "BelkinOReilly2009",
   # Check arguments
   checkArgs_df_array(x = x)
 
-  output <- lapply(seq(dim(x)[3]), function(i, ...) detectFronts(x = x[,,i], ...),
-                   method = method, intermediate = intermediate, checkPrevs = FALSE, ...)
+  output <- lapply(X = seq(dim(x)[3]),
+                   FUN = function(i, ...) detectFronts(x = x[,,i], ...),
+                   method = method, intermediate = intermediate,
+                   checkPrevs = FALSE, ...)
 
   if(isTRUE(intermediate)){
     index <- names(output[[1]])
-    output <- lapply(seq_along(index), function(x) abind(lapply(output, "[[", x), along = 3))
+    output <- lapply(X = seq_along(index),
+                     FUN = function(x) abind(lapply(output, "[[", x), along = 3))
 
     names(output) <- index
   }else{
     output <- abind(output, along = 3)
+
+    dimnames(output) <- dimnames(x)
   }
 
   output